[Chore #162828176] clean up code
[Finishes #162828176]
