+++ .travis.yml
@@ -4,7 +4,6 @@ node_js:
 cache:
   directories:
     - "node_modules"
-
 after_success:
   - npm run coverage
