+++ Procfile
@@ -1,1 +1,1 @@
-web: npm run build && node dist/app.js 
+web: npm run start:prod
