--- server/controllers/db/dateUtils.js
@@ -1,4 +1,0 @@
-/**
-* @returns {string} the current date in local String format
-*/
-export const getTimeString = () => (new Date()).toString();
